change menu vue page
