+++ src/views/Tab2.vue
@@ -12,17 +12,50 @@
         </ion-toolbar>
       </ion-header>
       
-      <ExploreContainer name="Tab 2 page" />
+       <ion-slides pager="true" :options="slideOpts">
+    <ion-slide>
+      <h1>메뉴 선택하기</h1>
+    </ion-slide>
+    <ion-slide>
+      <h1>빵 선택하기</h1>
+
+      <div class="box-container">
+        <div class="box">
+          
+        </div>
+        <div class="box">
+          
+        </div>
+        <div class="box">
+          
+        </div>
+      </div>
+    </ion-slide>
+    <ion-slide>
+      <h1>Slide 3</h1>
+    </ion-slide>
+  </ion-slides>
+  <ExploreContainer name="Tab 2 page" />
     </ion-content>
   </ion-page>
 </template>
 
 <script lang="ts">
-import { IonPage, IonHeader, IonToolbar, IonTitle, IonContent } from '@ionic/vue';
+import { IonPage, IonHeader, IonToolbar, IonTitle, IonContent, 
+IonSlides, IonSlide } from '@ionic/vue';
 import ExploreContainer from '@/components/ExploreContainer.vue';
 
 export default  {
   name: 'Tab2',
-  components: { ExploreContainer, IonHeader, IonToolbar, IonTitle, IonContent, IonPage }
+  components: { ExploreContainer, IonHeader, IonToolbar, IonTitle, IonContent, IonPage, 
+  IonSlides, IonSlide },
+  setup() {
+    // Optional parameters to pass to the swiper instance. See http://idangero.us/swiper/api/ for valid options.
+    const slideOpts = {
+      initialSlide: 1,
+      speed: 400
+    };
+    return { slideOpts }
+  }
 }
 </script>
